+++ Plugin/src/CodeRebirthInputs.cs
@@ -18,11 +18,8 @@ public class IngameKeybinds : LcInputActions {
     [InputAction("<Keyboard>/space", Name = "HoverUp")]
     public InputAction HoverUp { get; set; }
 
-    [InputAction("<Keyboard>/leftShift", Name = "DropHoverboard")]
+    [InputAction("<Keyboard>/c", Name = "DropHoverboard")]
     public InputAction DropHoverboard { get; set; }
-
-    [InputAction("<Keyboard>/p", Name = "TurnOnHoverboard")]
-    public InputAction TurnOnHoverboard { get; set; }
 
     [InputAction("<Keyboard>/f", Name = "SwitchMode")]
     public InputAction SwitchMode { get; set; }