+++ Plugin/src/ItemStuff/MaterialVariants.cs
@@ -10,29 +10,30 @@ namespace CodeRebirth.ItemStuff;
 public class MaterialVariants : NetworkBehaviour {
     [SerializeField]
     [Tooltip("The item data of the scrap.")]
-    Item itemData;
-        
+    private Item itemData;
+
     [Space(5f)]
     [SerializeField]
     [Tooltip("The mesh renderers to change the material of. This will use the first material in the array.")]
-    Renderer[] renderers;
+    private Renderer[] renderers;
 
     [FormerlySerializedAs("ChangeScanNodeText")]
     [Space(5f)]
-    bool changeScanNodeText;
+    private bool changeScanNodeText;
+    
     [SerializeField]
     [Tooltip("The text to change to when the material is changed.")]
-    string[] scanNodeText;
+    private string[] scanNodeText;
 
     [Space(5f)]
     [SerializeField]
     [Tooltip("The scan node properties to change the text of.")]
-    ScanNodeProperties scanNodeProperties;
+    private ScanNodeProperties scanNodeProperties;
 
     [Space(5f)]
     [SerializeField]
     [Tooltip("The currently saved material variant.")]
-    int savedMaterialVariant = -1;
+    private int savedMaterialVariant = -1;
 
     public override void OnNetworkSpawn() =>
         SetRendererServerRpc();