+++ Plugin/src/ScrapStuff/EpicAxe.cs
@@ -1,5 +1,0 @@
-
-
-namespace CodeRebirth.ScrapStuff;
-public class EpicAxe : Shovel { // Added for potential future implementations
-}