+++ Plugin/src/ScrapStuff/Hoverboard.cs
@@ -9,94 +9,90 @@ using UnityEngine.InputSystem.Controls;
 using UnityEngine.InputSystem;
 using UnityEngine.InputSystem.Interactions;
 using System.Linq;
+using System.Diagnostics.Tracing;
+using Unity.Netcode.Samples;
 
 namespace CodeRebirth.ScrapStuff;
 public class Hoverboard : GrabbableObject, IHittable
 {
-    Rigidbody hb;
+    public Rigidbody hb;
     public InteractTrigger trigger;
     private bool turnedOn = false;
     public float mult;
-    public float turnTorque;
     public GameObject hoverboardSeat;
-    public BoxCollider playerRidingCollider;
-    public PlayerControllerB playerControlling;
     public Transform[] anchors = new Transform[4];
-    public RaycastHit[] hits = new RaycastHit[4];
+    private PlayerControllerB playerControlling;
+    private RaycastHit[] hits = new RaycastHit[4];
     private bool _isHoverForwardHeld = false;
     private bool jumpCooldown = true;
-
+    public enum HoverboardMode {
+        None,
+        Held,
+        Mounted,
+    }
     public enum HoverboardTypes {
         Regular
+        // Eventually wanna make other types of hoverboards
     }
+    private HoverboardMode hoverboardMode;
     public override void Start()
     {
-        Plugin.InputActionsInstance.HoverForward.performed += OnHoverForward;
-        Plugin.InputActionsInstance.HoverForward.canceled += OnHoverForward;
-        Plugin.InputActionsInstance.HoverLeft.performed += MovementHandler;
-        Plugin.InputActionsInstance.HoverRight.performed += MovementHandler;
-        Plugin.InputActionsInstance.HoverBackward.performed += MovementHandler;
-        Plugin.InputActionsInstance.HoverForward.performed += MovementHandler;
-        Plugin.InputActionsInstance.HoverUp.performed += MovementHandler;
-        Plugin.InputActionsInstance.SwitchMode.performed += ModeHandler;
-
-        hb = GetComponent<Rigidbody>();
-        trigger = GetComponent<InteractTrigger>();
-        trigger.onInteract.AddListener(OnInteract);
+        StartBaseImportant();
+        if (IsHost) {
+            SetHoverboardStateClientRpc(0);
+        } else {
+            SetHoverboardStateServerRpc(0);
+        }
     }
     public void ModeHandler(InputAction.CallbackContext context) {
+        if (hoverboardMode == HoverboardMode.None || playerControlling != GameNetworkManager.Instance.localPlayerController) return;
         var btn = (ButtonControl)context.control;
         if (btn.wasPressedThisFrame)
         {
-            trigger.interactable = !trigger.interactable;
-            trigger.holdInteraction = trigger.interactable;
+            if (hoverboardMode == HoverboardMode.Mounted) {
+                if (IsHost) {
+                    SetHoverboardStateClientRpc(1);
+                } else {
+                    SetHoverboardStateServerRpc(1);
+                }
+                HandleToolTips();
+                SwitchModeExtension(true);
+            } else if (hoverboardMode == HoverboardMode.Held) {
+                if (IsHost) {
+                    SetHoverboardStateClientRpc(2);
+                } else {
+                    SetHoverboardStateServerRpc(2);
+                }
+                HandleToolTips();
+                SwitchModeExtension(false);
+            }
+            Plugin.InputActionsInstance.SwitchMode.performed += ModeHandler;
         }
     }
     public void OnInteract(PlayerControllerB player) {
-        if (trigger.holdInteraction) {
-            StartCoroutine(HandleGrabbing(this, player));
-            return;
+        if (hoverboardMode == HoverboardMode.None) {
+            SwitchModeExtension(false);
+            Plugin.InputActionsInstance.SwitchMode.performed += ModeHandler;
+            if (IsHost) {
+                SetTargetClientRpc(Array.IndexOf(StartOfRound.Instance.allPlayerScripts, player));
+                SetHoverboardStateClientRpc(2);
+            } else {
+                SetTargetServerRpc(Array.IndexOf(StartOfRound.Instance.allPlayerScripts, player));
+                SetHoverboardStateServerRpc(2);
+            }
+            HandleToolTips();
         }
-        if (!IsHost) SetTargetServerRpc(Array.IndexOf(StartOfRound.Instance.allPlayerScripts, player));
-        else SetTargetClientRpc(Array.IndexOf(StartOfRound.Instance.allPlayerScripts, player));
-        trigger.interactable = false;
     }
 
     public void OnHoverForward(InputAction.CallbackContext context) {
+        if (GameNetworkManager.Instance.localPlayerController != playerControlling) return;
         var btn = (ButtonControl)context.control;
         if (btn.wasPressedThisFrame)
-        {
-            // Button has started being held
             _isHoverForwardHeld = true;
-        }
-
-        if (btn.wasReleasedThisFrame)
-        {
-            // Button was released
+        else if (btn.wasReleasedThisFrame)
             _isHoverForwardHeld = false;
-        }
     }
 
-    public IEnumerator HandleGrabbing(GrabbableObject _obj, PlayerControllerB player) {
-        yield return new WaitForEndOfFrame();
-        trigger.interactable = false;
-        trigger.holdInteraction = false;
-        trigger.enabled = false;
-        _obj.InteractItem();
-        if(GameNetworkManager.Instance.localPlayerController.FirstEmptyItemSlot() == -1){
-            Plugin.Logger.LogInfo("GiveItemToPlayer: Could not grab item, inventory full!");
-            yield break;
-        }
-        player.twoHanded = _obj.itemProperties.twoHanded;
-        player.carryWeight += Mathf.Clamp(_obj.itemProperties.weight - 1f, 0f, 10f);
-        player.grabbedObjectValidated = true;
-        player.GrabObjectServerRpc(_obj.NetworkObject);
-        _obj.GrabItemOnClient();
-        _obj.parentObject = player.localItemHolder;
-        player.isHoldingObject = true;
-        _obj.hasBeenHeld = true;
-        _obj.EnablePhysics(false);
-    }
     public void MovementHandler(InputAction.CallbackContext context) {
         var btn = (ButtonControl)context.control;
         if (btn.wasPressedThisFrame)
@@ -104,64 +100,60 @@ public class Hoverboard : GrabbableObject, IHittable
             HandleMovement();
         }
     }
-    public override void DiscardItem() {
-        base.DiscardItem();
-        this.transform.position = playerHeldBy.transform.position + transform.forward * 1.5f;
-        trigger.enabled = true;
-        trigger.interactable = true;
-    }
     public void FixedUpdate()
     {
-        if (playerControlling != null) {
-            playerControlling.transform.position = hoverboardSeat.transform.position;
-            Quaternion playerRotation = playerControlling.transform.rotation;
-            Quaternion rotationOffset = Quaternion.Euler(0, -90, 0); // 90 degrees to the left around the y-axis
-            this.transform.rotation = playerRotation * rotationOffset;
-            playerControlling.ResetFallGravity();
-        }
-        if (_isHoverForwardHeld && playerControlling == GameNetworkManager.Instance.localPlayerController)
-            hb.AddForce(Vector3.zero + transform.right * 40f * (turnedOn ? 1f : 0.1f), ForceMode.Acceleration);
-        if (!turnedOn) return;
-        for (int i = 0; i < 4; i++)
-            ApplyForce(anchors[i], hits[i]);
+        if (!IsHost || hoverboardMode != HoverboardMode.Mounted || playerControlling == null) return;
+        ApplyForcesForHoverboardClientRpc();
     }
 
     public override void Update()
     {
         base.Update();
+        HandleDropping();
+        HandleToolTips();
+        if (!IsHost) return;
+        PlayerControllerB realPlayer = StartOfRound.Instance.allPlayerScripts.FirstOrDefault();
+        if (Vector3.Distance(transform.position, StartOfRound.Instance.shipBounds.transform.position) < 15 && !isInShipRoom) {
+            this.transform.SetParent(realPlayer.playersManager.elevatorTransform, true);
+            isInShipRoom = true;
+            isInElevator = true;
+        } else if (Vector3.Distance(transform.position, StartOfRound.Instance.shipBounds.transform.position) >= 15 && isInShipRoom) {
+            this.transform.SetParent(realPlayer.playersManager.propsContainer, true);
+            isInShipRoom = false;
+            isInElevator = false;
+        }
         if (playerControlling == null) return;
-        if (playerControlling.inAnimationWithEnemy || playerControlling.inSpecialInteractAnimation || playerControlling.isClimbingLadder) {
-            DropHoverboard(false, true);
+        if (playerControlling.inAnimationWithEnemy || playerControlling.inSpecialInteractAnimation || (hoverboardMode == HoverboardMode.Mounted && playerControlling.isClimbingLadder)) {
+            DropHoverboard(true);
             return;
         }
-        HandleDropping();
-        HandleTurningOnOrOff();
+        SetHoverboardPositionClientRpc();
     }
 
-    public void HandleTurningOnOrOff() {
-        if (!Plugin.InputActionsInstance.TurnOnHoverboard.triggered) return;
-        turnedOn = !turnedOn;
-    }
     public void HandleDropping() {
         if (!Plugin.InputActionsInstance.DropHoverboard.triggered) return;
-        DropHoverboard(turnedOn, true);
+        DropHoverboard(true);
     }
-    public void DropHoverboard(bool keepOn = true, bool stillWorks = true) {
-        if (!IsHost) SetTargetServerRpc(-1);
-        else SetTargetClientRpc(-1);
-        turnedOn = keepOn;
+    public void DropHoverboard(bool stillWorks = true) {
+        if (IsHost) {
+            SetHoverboardStateClientRpc(0);
+            SetTargetClientRpc(-1);
+        } else {
+            SetHoverboardStateServerRpc(0);
+            SetTargetServerRpc(-1);
+        }
+        HandleToolTips();
+        SwitchModeExtension(true);
+        Plugin.InputActionsInstance.SwitchMode.performed -= ModeHandler;
         trigger.interactable = stillWorks;
     }
-    public override void LateUpdate() {
-        base.LateUpdate();
-    }
+
     private void HandleMovement() // the reason these transform.forward and transform.right seemingly don't match with the buttons is because the exported hoverboard is kinda fucked... oh well.
     {
-        if (playerControlling == null) return;
-        if (GameNetworkManager.Instance.localPlayerController != playerControlling) return;
+        if (playerControlling == null || GameNetworkManager.Instance.localPlayerController != playerControlling) return;
         Vector3 forceDirection = Vector3.zero;
         float moveForce = 100f;
-
+        
         if (Plugin.InputActionsInstance.HoverLeft.WasPressedThisFrame())
             forceDirection += transform.forward;
 
@@ -174,7 +166,7 @@ public class Hoverboard : GrabbableObject, IHittable
         if (Plugin.InputActionsInstance.HoverForward.WasPressedThisFrame())
             forceDirection += transform.right;
 
-        if (Plugin.InputActionsInstance.HoverUp.WasPressedThisFrame() && turnedOn && jumpCooldown) {
+        if (Plugin.InputActionsInstance.HoverUp.WasPressedThisFrame() && jumpCooldown) {
             jumpCooldown = false;
             forceDirection += transform.up;
             moveForce = 1000f;
@@ -182,18 +174,11 @@ public class Hoverboard : GrabbableObject, IHittable
         }
 
         if (forceDirection == Vector3.zero) return;
-        Plugin.Logger.LogInfo("moveForce: " + moveForce);
-        Plugin.Logger.LogInfo("forceDirection: " + forceDirection);
-        if (!turnedOn) 
-            moveForce *= 0.1f;
         hb.AddForce(forceDirection * moveForce, ForceMode.Acceleration);
     }
     public IEnumerator JumpTimerStart() {
         yield return new WaitForSeconds(2f);
         jumpCooldown = true;
-    }
-    public override void FallWithCurve() {
-        return;
     }
     
     public void ApplyForce(Transform anchor, RaycastHit hit)
@@ -217,6 +202,21 @@ public class Hoverboard : GrabbableObject, IHittable
         SetTargetClientRpc(PlayerID);
     }
     [ClientRpc]
+    internal void ApplyForcesForHoverboardClientRpc() {
+        if (Vector3.Distance(transform.position, playerControlling.transform.position) > 5f) {
+            DropHoverboard(true);
+            return;
+        }
+        if (_isHoverForwardHeld && playerControlling == GameNetworkManager.Instance.localPlayerController) {
+            hb.AddForce(Vector3.zero + transform.right * 40f, ForceMode.Acceleration);
+        }
+        if (turnedOn) {
+            for (int i = 0; i < 4; i++) {
+                ApplyForce(anchors[i], hits[i]);
+            }
+        }
+    }
+    [ClientRpc]
     internal void SetTargetClientRpc(int PlayerID) {
         if (PlayerID == -1) {
             playerControlling = null;
@@ -231,5 +231,147 @@ public class Hoverboard : GrabbableObject, IHittable
         playerControlling = StartOfRound.Instance.allPlayerScripts[PlayerID];
         // playerRidingCollider.enabled = true;
         Plugin.Logger.LogInfo($"{this} setting target to: {playerControlling.playerUsername}");
+    }
+    [ClientRpc]
+    internal void SetHoverboardPositionClientRpc() {
+        Quaternion playerRotation = playerControlling.transform.rotation;
+        Quaternion rotationOffset;
+        if (hoverboardMode == HoverboardMode.Held) {
+            // Position the hoverboard to the right side of the player
+            this.transform.position = playerControlling.transform.position + playerControlling.transform.right * 0.7f + playerControlling.transform.up * 1f;
+
+            // Make the hoverboard face upwards with its up vector pointing to the player's right hand
+            rotationOffset = Quaternion.Euler(180, 180, -90); // Adjust to match correct facing direction
+            this.transform.rotation = playerRotation * rotationOffset;
+            return;
+        }
+
+        if (hoverboardMode == HoverboardMode.Mounted) {
+            playerControlling.transform.position = hoverboardSeat.transform.position;
+            rotationOffset = Quaternion.Euler(0, -90, 0); // 90 degrees to the left around the y-axis
+            this.transform.rotation = playerRotation * rotationOffset;
+            playerControlling.ResetFallGravity();
+        }
+    }
+    [ServerRpc(RequireOwnership = false)]
+    internal void SetHoverboardStateServerRpc(int state) {
+        SetHoverboardStateClientRpc(state);
+    }
+    [ClientRpc]
+    internal void SetHoverboardStateClientRpc(int state) {
+        switch(state) {
+            case 0:
+                hoverboardMode = HoverboardMode.None;
+                turnedOn = false;
+                _isHoverForwardHeld = false;
+                hb.useGravity = true;
+                hb.isKinematic = false;
+                trigger.interactable = true;
+                break;
+            case 1:
+                hoverboardMode = HoverboardMode.Held;
+                _isHoverForwardHeld = false;
+                turnedOn = false;
+                hb.useGravity = false;
+                hb.isKinematic = true;
+                break;
+            case 2:
+                hoverboardMode = HoverboardMode.Mounted;
+                turnedOn = true;
+                hb.useGravity = true;
+                hb.isKinematic = false;
+                trigger.interactable = false;
+                break;
+            default:
+                hoverboardMode = HoverboardMode.None;
+                _isHoverForwardHeld = false;
+                turnedOn = false;
+                hb.useGravity = true;
+                hb.isKinematic = false;
+                trigger.interactable = true;
+                Plugin.Logger.LogInfo("Invalid state!");
+                break;
+        }
+    }
+    public override void FallWithCurve() {
+        return;
+    }
+    public void StartBaseImportant() {
+        this.propColliders = base.gameObject.GetComponentsInChildren<Collider>();
+		this.originalScale = base.transform.localScale;
+		if (this.itemProperties.itemSpawnsOnGround)
+		{
+			this.startFallingPosition = base.transform.position;
+			if (base.transform.parent != null)
+			{
+				this.startFallingPosition = base.transform.parent.InverseTransformPoint(this.startFallingPosition);
+			}
+			this.FallToGround(false);
+		}
+		else
+		{
+			this.hasHitGround = true;
+			this.reachedFloorTarget = true;
+			this.targetFloorPosition = base.transform.localPosition;
+		}
+		if (this.itemProperties.isScrap)
+		{
+			this.hasHitGround = true;
+		}
+		if (this.itemProperties.isScrap && RoundManager.Instance.mapPropsContainer != null)
+		{
+			this.radarIcon = Instantiate<GameObject>(StartOfRound.Instance.itemRadarIconPrefab, RoundManager.Instance.mapPropsContainer.transform).transform;
+		}
+		if (!this.itemProperties.isScrap)
+		{
+			HoarderBugAI.grabbableObjectsInMap.Add(base.gameObject);
+		}
+		MeshRenderer[] componentsInChildren = base.gameObject.GetComponentsInChildren<MeshRenderer>();
+		for (int i = 0; i < componentsInChildren.Length; i++)
+		{
+			componentsInChildren[i].renderingLayerMask = 1U;
+		}
+		SkinnedMeshRenderer[] componentsInChildren2 = base.gameObject.GetComponentsInChildren<SkinnedMeshRenderer>();
+		for (int j = 0; j < componentsInChildren2.Length; j++)
+		{
+			componentsInChildren2[j].renderingLayerMask = 1U;
+		}
+        trigger.onInteract.AddListener(OnInteract);
+    }
+    public void SwitchModeExtension(bool SwitchingOff) {
+        if (SwitchingOff) {
+            Plugin.InputActionsInstance.HoverForward.performed -= OnHoverForward;
+            Plugin.InputActionsInstance.HoverForward.canceled -= OnHoverForward;
+            Plugin.InputActionsInstance.HoverLeft.performed -= MovementHandler;
+            Plugin.InputActionsInstance.HoverRight.performed -= MovementHandler;
+            Plugin.InputActionsInstance.HoverBackward.performed -= MovementHandler;
+            Plugin.InputActionsInstance.HoverForward.performed -= MovementHandler;
+            Plugin.InputActionsInstance.HoverUp.performed -= MovementHandler;
+        } else {
+            Plugin.InputActionsInstance.HoverForward.performed += OnHoverForward;
+            Plugin.InputActionsInstance.HoverForward.canceled += OnHoverForward;
+            Plugin.InputActionsInstance.HoverLeft.performed += MovementHandler;
+            Plugin.InputActionsInstance.HoverRight.performed += MovementHandler;
+            Plugin.InputActionsInstance.HoverBackward.performed += MovementHandler;
+            Plugin.InputActionsInstance.HoverForward.performed += MovementHandler;
+            Plugin.InputActionsInstance.HoverUp.performed += MovementHandler;
+        }
+    }
+    public void HandleToolTips() {
+        if (playerControlling == null || GameNetworkManager.Instance.localPlayerController != playerControlling) return;
+        HUDManager.Instance.ClearControlTips();
+        if (hoverboardMode == HoverboardMode.Mounted) {
+            HUDManager.Instance.ChangeControlTipMultiple([
+                $"Dismount Hoverboard : [{Plugin.InputActionsInstance.DropHoverboard.GetBindingDisplayString().Split(" ")[0]}]",
+                $"Move Hoverboard : [{Plugin.InputActionsInstance.HoverForward.GetBindingDisplayString().Split(" ")[0]}][{Plugin.InputActionsInstance.HoverLeft.GetBindingDisplayString().Split(" ")[0]}][{Plugin.InputActionsInstance.HoverBackward.GetBindingDisplayString().Split(" ")[0]}][{Plugin.InputActionsInstance.HoverRight.GetBindingDisplayString().Split(" ")[0]}]",
+                $"Up Boost : [{Plugin.InputActionsInstance.HoverUp.GetBindingDisplayString().Split(" ")[0]}]",
+                $"Switch Mode (Mounted) : [{Plugin.InputActionsInstance.SwitchMode.GetBindingDisplayString().Split(" ")[0]}]",
+            ]);
+        } else if (hoverboardMode == HoverboardMode.Held) {
+            HUDManager.Instance.ChangeControlTipMultiple([
+                $"Drop Hoverboard : [{Plugin.InputActionsInstance.DropHoverboard.GetBindingDisplayString().Split(" ")[0]}]",
+                $"Switch Mode (Held) : [{Plugin.InputActionsInstance.SwitchMode.GetBindingDisplayString().Split(" ")[0]}]",
+            ]);
+        }
     }
 }