+++ Plugin/src/Util/Extensions/ShovelExtensions.cs
@@ -1,0 +1,16 @@
+using System;
+using System.Diagnostics;
+using BepInEx.Logging;
+using UnityEngine;
+
+namespace CodeRebirth.Util.Extensions;
+
+public static class ShovelExtensions {
+	public static int CriticalHit(int force, System.Random random, int critChance) {
+		if (random.Next(0, 100) < Math.Clamp(critChance, 0, 99)) {
+            Plugin.Logger.LogInfo("Critical Hit!");
+            return force * 2;
+        }
+        return force;
+	}
+}