+++ Plugin/src/WeaponStuff/CodeRebirthWeapons.cs
@@ -1,0 +1,11 @@
+namespace CodeRebirth.WeaponStuff;
+public class CodeRebirthWeapons : Shovel
+{
+    public int defaultForce = 0;
+    public bool critPossible = true;
+    public int critChance = 25;
+
+    public void Awake() {
+        defaultForce = shovelHitForce;
+    }
+}